split up the smoke tests
