+++ packages/compass-e2e-tests/helpers/commands/click-visible.ts
@@ -5,6 +5,7 @@ export async function clickVisible(
   selector: string
 ): Promise<void> {
   const element = await browser.$(selector);
+  await element.scrollIntoView();
   await element.waitForDisplayed();
   await browser.waitForAnimations(selector);
   await element.click();
