+++ packages/compass-e2e-tests/tests/collection-indexes-tab.test.ts
@@ -1,0 +1,46 @@
+import chai from 'chai';
+import type { CompassBrowser } from '../helpers/compass-browser';
+import {
+  beforeTests,
+  afterTests,
+  afterTest,
+  Compass,
+} from '../helpers/compass';
+import * as Selectors from '../helpers/selectors';
+
+const { expect } = chai;
+
+describe('Collection indexes tab', function () {
+  let compass: Compass;
+  let browser: CompassBrowser;
+
+  before(async function () {
+    compass = await beforeTests();
+    browser = compass.browser;
+
+    await browser.connectWithConnectionString('mongodb://localhost:27018/test');
+
+    await browser.navigateToCollectionTab('test', 'numbers', 'Indexes');
+  });
+
+  after(async function () {
+    await afterTests(compass);
+  });
+
+  afterEach(async function () {
+    await afterTest(compass, this.currentTest);
+  });
+
+  it('lists indexes', async function () {
+    const element = await browser.$(Selectors.IndexList);
+    await element.waitForDisplayed();
+
+    const indexes = await browser.$$(Selectors.IndexComponent);
+    expect(indexes).to.have.lengthOf(1);
+
+    const nameColumnNameElement = await browser.$(Selectors.NameColumnName);
+    expect(await nameColumnNameElement.getText()).to.equal('_id_');
+  });
+
+  it('supports creating and dropping indexes');
+});
