+++ packages/compass-e2e-tests/tests/database-collections-tab.test.ts
@@ -1,0 +1,42 @@
+import chai from 'chai';
+import type { CompassBrowser } from '../helpers/compass-browser';
+import {
+  beforeTests,
+  afterTests,
+  afterTest,
+  Compass,
+} from '../helpers/compass';
+import * as Selectors from '../helpers/selectors';
+
+const { expect } = chai;
+
+describe('Database collections tab', function () {
+  let compass: Compass;
+  let browser: CompassBrowser;
+
+  before(async function () {
+    compass = await beforeTests();
+    browser = compass.browser;
+
+    await browser.connectWithConnectionString('mongodb://localhost:27018/test');
+
+    await browser.navigateToDatabaseTab('test', 'Collections');
+  });
+
+  after(async function () {
+    await afterTests(compass);
+  });
+
+  afterEach(async function () {
+    await afterTest(compass, this.currentTest);
+  });
+
+  it('contains a list of collections', async function () {
+    expect(await browser.existsEventually(Selectors.CollectionsGrid)).to.eq(
+      true
+    );
+  });
+
+  // capped and not capped
+  it('can create a collection and drop it');
+});
